+++ dimos/control/examples/twist_base_keyboard_teleop.py
@@ -35,7 +35,7 @@ from __future__ import annotations
 
 from dimos.control.blueprints.mobile import coordinator_mock_twist_base
 from dimos.core.coordination.module_coordinator import ModuleCoordinator
-from dimos.robot.unitree.keyboard_teleop import KeyboardTeleop
+from dimos.teleop.keyboard.twist_keyboard_teleop import KeyboardTeleop
 
 
 def main() -> None:
