+++ dimos/robot/unitree/g1/blueprints/agentic/unitree_g1_full.py
@@ -18,7 +18,7 @@
 from dimos.core.coordination.blueprints import autoconnect
 from dimos.robot.unitree.g1.blueprints.agentic._agentic_skills import _agentic_skills
 from dimos.robot.unitree.g1.blueprints.perceptive.unitree_g1_shm import unitree_g1_shm
-from dimos.robot.unitree.keyboard_teleop import KeyboardTeleop
+from dimos.teleop.keyboard.twist_keyboard_teleop import KeyboardTeleop
 
 unitree_g1_full = autoconnect(
     unitree_g1_shm,
