+++ dimos/robot/unitree/g1/blueprints/basic/unitree_g1_joystick.py
@@ -17,7 +17,7 @@
 
 from dimos.core.coordination.blueprints import autoconnect
 from dimos.robot.unitree.g1.blueprints.basic.unitree_g1_basic import unitree_g1_basic
-from dimos.robot.unitree.keyboard_teleop import KeyboardTeleop
+from dimos.teleop.keyboard.twist_keyboard_teleop import KeyboardTeleop
 
 unitree_g1_joystick = autoconnect(
     unitree_g1_basic,
